audioboard.c:DACTest sound.def,compile_audio.py: macro-based audio file index test runtime baud rate computation lib/audio_out.c:sample rate to 12kHz (need to factor into a .def file)
git-svn-id: https://svn.circlemud.org/svn/rocket/trunk@500 46170a15-b891-4bf7-af97-fd374ce181e0

+++ src/lib/audio_streamer.c
@@ -80,7 +80,7 @@ static inline uint8_t _ad_decode_ulaw(uint8_t ulaw)
 	return _ad_pcm16s_to_pcm8u(_ad_ulaw2linear(ulaw));
 }
 
-void _ad_decode_ulaw_buf(uint8_t *dst, uint8_t *src, uint8_t len)
+void _ad_decode_ulaw_buf(uint8_t *dst, uint8_t *src, uint16_t len)
 {
 	uint8_t *end = src+len;
 	for (; src<end; src++, dst++)
